add UI and behavior for aggregating results from multiple servers, Co-Authored-By: termermc <termer@termer.net>
